SubglobalSensitivityAnalysis

Documentation for SubglobalSensitivityAnalysis.

SubglobalSensitivityAnalysis.calculate_parbounds

SubglobalSensitivityAnalysis.calculate_parbounds!

SubglobalSensitivityAnalysis.compute_cp_design_matrix

SubglobalSensitivityAnalysis.compute_sobol_indices

SubglobalSensitivityAnalysis.estimate_subglobal_sobol_indices

SubglobalSensitivityAnalysis.fit_distributions

SubglobalSensitivityAnalysis.fit_distributions!

SubglobalSensitivityAnalysis.set_reference_parameters!

SubglobalSensitivityAnalysis.transform_cp_design_to_quantiles

SubglobalSensitivityAnalysis.calculate_parbounds! — Method

calculate_parbounds(dist, x; δ_cp = 0.1 )

compute the values at quantiles ±δcp around x with δcp difference in the cumulated probability. The quantiles are constrained to not extend beyond min_quant and max_quant.

It returns a NamedTuple of

sens_lowe: lower quantile

sens_upper: upper quantile

cp_ref: cdf of x (cumulative distribution function, i.e. p-value)

cpsenslower: cdf of lower quantile

cpsensupper: cdf of upper quantile

A wider distribution prior distribution will result in a wider intervals.

The DataFrame variant assumes x as column :ref to be present and adds/modifies output columns (named as above outputs).

SubglobalSensitivityAnalysis.calculate_parbounds — Method

calculate_parbounds(dist, x; δ_cp = 0.1 )

compute the values at quantiles ±δcp around x with δcp difference in the cumulated probability. The quantiles are constrained to not extend beyond min_quant and max_quant.

It returns a NamedTuple of

sens_lowe: lower quantile

sens_upper: upper quantile

cp_ref: cdf of x (cumulative distribution function, i.e. p-value)

cpsenslower: cdf of lower quantile

cpsensupper: cdf of upper quantile

A wider distribution prior distribution will result in a wider intervals.

The DataFrame variant assumes x as column :ref to be present and adds/modifies output columns (named as above outputs).

SubglobalSensitivityAnalysis.compute_cp_design_matrix — Method

compute_cp_design_matrix(df_dist, names_opt, N; path_sens_object=tempname()*".rds")

Compute the design matrix for two symples of size N from cumulative probability ranges. Returns a NamedTuple:

cpdesign: Matrix (nrow x n_param) for wich output needs to be computed

pathsensobject: path to the file that stores the R sensitivity object

SubglobalSensitivityAnalysis.compute_sobol_indices — Method

compute_sobol_indices(y, path_sens_object, par_names)

Tell the results, y,to sensitivity object in R, deserialized from path_sens_object and compute first order and total SOBOL effects and its uncertainty.

Returns a DataFrame with columns

par: parameter name from parnames - should match `dfcfopt.par` from computecpdesign_matrix

index: which one of the SOBOL-indices, :first_order or :total

value: the estimate

cf95lower and cf95upper: estimates of the 95% confidence interval

SubglobalSensitivityAnalysis.estimate_subglobal_sobol_indices — Method

estimate_regional_sobol_indices(f, parmsModeUpperRows, p0; 
    n_sample = 500, δ_cp = 0.1)

Estimate the Sobol sensitivity indices for a subset of the global space around parameter vector p0.

The region subspace to sample is determined by an area in the cumulative probability function, specifically for parameter i: cdf(p0) ± δ_cp, Samples are drawn from this region in the cdf and converted back to quantiles at the parameter scale.

Sobol indices are estimated using the method of Touati (2016), which has a total cost of $(p+2)×n$, where p is the number of parameters and n is the number of samples in each of the two random parameter samples.

Arguments

f: a function to compute a set of results, whose sensitivity is to be inspected,

from parametes (p1, p2, ...) -> NamedTuple{NTuple{N,NT}} where NT <: Number, for example fsens = (a,b) -> (;target1 = a + b -1, target2 = a + b -0.5).

parmsModeUpperRows: a Vector of Tuples of the form (:par_name, Distribution, mode, 95%_quantile) where Distribution is a non-parameterized Object from Distributions.jl such as LogNormal

p0: the parameter around which the samples are drawn.

Optional

n_sample = 500: the number of parameter-vectors in each of the two samples

δcp = 0.1: the range around cdf(p0i) to sample.

targets: a NTuple{Symbol} of subset of the outputs of f, to constrain the computation to specific outputs.

min_quant=0.005 and max_quant=0.995: to constrain the range of cumulative probabilities when parameters are near the ends of the distribution.

Return value

A DataFrame with columns

par: parameter name

index: which one of the SOBOL-indices, :first_order or :total

value: the estimate

cf95lower and cf95upper: estimates of the 95% confidence interval

target: the result, for which the sensitivity has been computed

SubglobalSensitivityAnalysis.fit_distributions! — Method

fit_distributions(tups)
fit_distributions!(df)

For each row, use fit a distribution of type dType to mode and upper quantile.

In the first variant, parameters are specified as a vector of tuples, which are converted to a DataFrame. A new column :dist with a concrete Distribution is added. The second variant modifies a DataFrame with corresponding input columns.

SubglobalSensitivityAnalysis.fit_distributions — Method

fit_distributions(tups)
fit_distributions!(df)

For each row, use fit a distribution of type dType to mode and upper quantile.

In the first variant, parameters are specified as a vector of tuples, which are converted to a DataFrame. A new column :dist with a concrete Distribution is added. The second variant modifies a DataFrame with corresponding input columns.

SubglobalSensitivityAnalysis.set_reference_parameters! — Method

set_reference_parameters!(df, par_dict)

Set the :ref column to given parameters for keys in par_dict matching column :par. Non-matching keys are set to missing.

SubglobalSensitivityAnalysis.transform_cp_design_to_quantiles — Method

transform_cp_design_to_quantiles(df_cfopt, cp_design)

Transform cumulative probabilities back to quantiles.