Internal methods

SubglobalSensitivityAnalysis.calculate_parbounds!

SubglobalSensitivityAnalysis.get_uniform_cp_sample

SubglobalSensitivityAnalysis.calculate_parbounds! — Method

calculate_parbounds(dist, x; δ_cp = 0.1 )

compute the values at quantiles ±δcp around x with δcp difference in the cumulated probability. The quantiles are constrained to not extend beyond min_quant and max_quant.

It returns a NamedTuple of

sens_lowe: lower quantile

sens_upper: upper quantile

cp_ref: cdf of x (cumulative distribution function, i.e. p-value)

cpsenslower: cdf of lower quantile

cpsensupper: cdf of upper quantile

A wider distribution prior distribution will result in a wider intervals.

The DataFrame variant assumes x as column :ref to be present and adds/modifies output columns (named as above outputs).

SubglobalSensitivityAnalysis.get_uniform_cp_sample — Method

get matrix (nsample, npar) with uniformly sampled in cumaltive p domain